Check DAG consistency only once the DAG is complete
DAG.add_vertex checks consistency at each call, which is costly and
creates superlinear complexity when adding tests: it easily takes 50s on
a powerful machine to load a testsuite with ~2600 tests. Check the
consistency only once the DAG is complete instead to have the same level
of checking, but with neglectible overhead.

TN: T527-050