AdaCoreLegacyTestDriver: fix working directory substitution
The path substitution for the test working directory is supposed (for
compatibility with legacy) to also replace characters that are not
space/single quote/double quote that precede the directory name, but
should not cross line separators, unlike the current behavior.

This commit makes this output refiner process each line separately to
make its implementation closer to the legacy scripts.

Fixes #56