Always enable "cross" support for testsuites
The name is misleading: always add support for --build/--host/--target
options. This is useful even when testing a native-only project, as it
allows the testsuite to be aware that it is testing a 32-bit project on
a 64-bit machine, which is useful for instance to adjust test control.

TN: TB30-002