AdaCoreLegacyTestDriver: stop altering shell scripts
The AdaCoreLegacyTestDriver assumes the content of shell scripts uses
CMD syntax instead of Bourne syntax and needs to be converted on
non-windows hosts. This is both counter-intuitive and wrong in the
case of the fixedbugs testsuite.
This patch changes this: testcases are now only converted if their
extension starts with ".cmd" and the host isn't windows or conversion
has been explicitly requested.

TN: TB25-014