Testsuite: enhance the API to ease adding inter-test dependencies
This commit introduces the FragmentData class to clarify what the
testsuite DAG contains (used to be a tuple) and to contain intermediate
data about the fragment: "simple" name (i.e. UID suffix), test driver
instance involved (not just a callback). It then also introduce the
adjust_dag_dependencies method so that testsuite have a chance to amend
predecessors in the testsuite DAG after all tests have been added but
before the scheduler starts to process it.

The goal is to make it easier for testsuites to create inter-test
dependencies: so far, only intra-test deps were really supported.

TN: U317-035