--generate-text-report: show logs in same order as in the summary
In the text reports summary, we display new failures, then already
detected failures, then fixed failures, etc. in that order because we
want to display the most important ("urgent to investigate") results
first. For the same reason, we should display test logs in the same
order (having a consistent order is also less confusing). This commit
implements this.

TN: U621-003