Introduce multi-processing for test fragment execution
Beyond a certain level of parallelism, Python's GIL contention is too
high to benefit from more processors. When we reach this level, it is
more interesting to use multiple processes to cancel the GIL contention.

This commit introduces what is necessary to run test fragments in
separate processes, and introduce a scheduler that does not use
multithreading in order to spawn and coordinate subprocesses, all for
better performances at high levels of parallelism.

This new mechanism is disabled by default, as it puts more requirements
on testsuites: all code run by test drivers must be importable from
subprocesses, test environments and results must be compatible with the
pickle module and return value propagation between tests is disabled
("previous_values" is always the empty dict).

TN: U614-006