--gaia-output: create the "discs" file in more cases
OptFileParse accepts both a string (comma-separated list of
discriminants) and a list of strings (the list of discriminants itself).
To be consistent with it, enhance the implementation of --gaia-output so
that it creates the "discs" file in both cases: before this commit, it
used to do it only when "self.env.discs" contained a list of strings.

TN: UC08-006