Don't walk into internal VCS directories when looking for tests
Added test, without the patch, would cause multiple failures when
running the tox command on e3-testsuite itself, e.g. from the TestBasic
in test_basics.py, which expects simple-tests directory to contain just
test1 and test2. With the patch this test provides coverage for added
code. There is no additional expected output.

We could add a similar test for the ".svn" directory, but not for
".git", because git doesn't repositories to contain ".git" directories.
For coverage and testing the "CVS" directory is enough.