Rework the handling of maximum consecutive failures
When the testsuite aborts because there are too many consecutive
failures, keep running the mainloop, but tune the test fragments so that
instead of actually running them, we push a SKIP test result instead.

This both fixes the suspiciously low test result count in the testsuite
report in case the testsuite is aborted, and makes the report include
all test results that are added to the report "in batch" (they were
capped otherwise).