Fix test filtering when not all test finders have dedicated directories
When one test finder TF1 guarantees dedicated directories for its tests
but not TF2, and when a testsuite using these two test finders is run
with a filtering pattern on the command line, e3.testsuite currently
does not use the pattern to filter the directories to probe.

This is expected for TF2: when tests do not have their own directories,
filtering is done using the test matcher only. However, the pattern
should still be matches against the test directory for TF1: most test
finders with dedicated dirs do not provide a test matcher, so we need to
rely on the test directory name to do the filtering.

This change corrects the filtering logic so that the directory name is
ignored only for tests found by TF2.