Add job to upload to package registry
The package registry currently builds the wheel itself with a
version including the date of the build like
<major>.<minor>.<build_date>. To fix an ordering issue between
"official" and intermediate releases, we now want the patch version
to be the number of commits since the last major/minor version
bump. We also want the build and upload to package registry to
be done by the pipeline of e3-testsuite itself.

The minor version of e3-testsuite has to be bumped to avoid
conflicts between the new versioning system and the old one.

Ref it/org/operation_support/iaas/issues#214