RunningStatus: do not store references to TestFragment instances
`RunningStatus` objects live as long as the testsuite instance. Prevent
them from storing references to `TestFragment` instances, so that
these objects can be disposed, as well as the associated resources.

In particular for multithreaded runs, `TestFragment` inherits
`e3.job.Job`, which contains a reference to a `threading.Thread`
objects, which itself contains a lock. On Windows, this creates a leak
of handles.